Handle Compilation Threads at checkpoint/restore

+++ runtime/compiler/control/CompilationThread.cpp
@@ -471,6 +471,12 @@ int32_t TR::CompilationInfo::computeDynamicDumbInlinerBytecodeSizeCutoff(TR::Opt
 // Must have compilation queue monitor in hand when calling this routine
 TR_YesNoMaybe TR::CompilationInfo::shouldActivateNewCompThread()
    {
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+   // If a checkpoint is in progress, don't activate any threads until the restore.
+   if (isCheckpointInProgress())
+      return TR_no;
+#endif
+
    // If further compilations have been disabled we could be in a dire situation, for example virtual memory could be
    // really low, there could be failures when attempting to allocate persistent memory, a compilation thread could
    // have crashed, etc. In such situations we never want to activate a new compilation driven by this API.
@@ -1126,6 +1132,10 @@ TR::CompilationInfo::CompilationInfo(J9JITConfig *jitConfig) :
 #if defined(J9VM_JIT_DYNAMIC_LOOP_TRANSFER)
    _dltMonitor = TR::Monitor::create("JIT-DLTmonitor");
 #endif
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+   _crMonitor = TR::Monitor::create("JIT-CheckpointRestoreMonitor");
+   _checkpointStatus = TR_CheckpointStatus::NO_CHECKPOINT_IN_PROGRESS;
+#endif
    _iprofilerBufferArrivalMonitor = TR::Monitor::create("JIT-IProfilerBufferArrivalMonitor");
    _classUnloadMonitor = TR::MonitorTable::get()->getClassUnloadMonitor(); // by this time this variable is initialized
                                              // TODO: hang these monitors to something persistent
@@ -1428,6 +1438,23 @@ void TR::CompilationInfo::releaseCompilationLock()
       releaseCompMonitor(0);
       }
    }
+
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+void TR::CompilationInfo::acquireCRMonitor()
+   {
+   getCRMonitor()->enter();
+   }
+
+void TR::CompilationInfo::releaseCRMonitor()
+   {
+   getCRMonitor()->exit();
+   }
+
+void TR::CompilationInfo::waitOnCRMonitor()
+   {
+   getCRMonitor()->wait();
+   }
+#endif
 
 TR::Monitor * TR::CompilationInfo::createLogMonitor()
    {
@@ -2428,13 +2455,8 @@ void TR::CompilationInfoPerThread::suspendCompilationThread()
 // This method may be called when we want all JIT activity to stop temporarily
 // It is called normally by a java thread
 //----------------------------------------------------------------------------
-void TR::CompilationInfo::suspendCompilationThread()
+void TR::CompilationInfo::suspendCompilationThread(bool purgeCompQueue)
    {
-   // if we compile on application thread, there is no compilation
-   // request queue and there is no compilation thread
-   // There can only be one request that is in progress. All the others
-   // will see the SUSPENDED state when they get the application monitor
-
    TR_ASSERT(_compilationMonitor, "Must have a compilation queue monitor\n");
    J9JavaVM   *vm       = _jitConfig->javaVM;
    J9VMThread *vmThread = vm->internalVMFunctions->currentVMThread(vm);
@@ -2463,7 +2485,7 @@ void TR::CompilationInfo::suspendCompilationThread()
             }
          }
       }
-   if (stoppedOneCompilationThread)
+   if (stoppedOneCompilationThread && purgeCompQueue)
       purgeMethodQueue(compilationSuspended);
    TR_ASSERT(_numCompThreadsActive == 0, "We must have all compilation threads suspended at this point\n");
    releaseCompMonitor(vmThread);
@@ -2596,6 +2618,93 @@ void TR::CompilationInfo::resumeCompilationThread()
    releaseCompMonitor(vmThread);
    }
 
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+void TR::CompilationInfo::prepareForCheckpoint()
+   {
+   J9JavaVM   *vm       = _jitConfig->javaVM;
+   J9VMThread *vmThread = vm->internalVMFunctions->currentVMThread(vm);
+
+   {
+   OMR::CriticalSection suspendCompThreadsForCheckpoint(getCompilationMonitor());
+
+   if (shouldCheckpointBeInterrupted())
+      return;
+
+   TR_ASSERT_FATAL(!isCheckpointInProgress(), "Checkpoint already in progress!\n");
+
+   /* Set the checkpoint in progress status. */
+   setCheckpointInProgress();
+
+   /* Inform compilation threads to suspend themselves. The compilation
+    * threads will complete any in-progress compilations first.
+    */
+   bool purgeMethodQueue = false;
+   suspendCompilationThread(purgeMethodQueue);
+
+   /* Wait until all compilation threads are suspended. */
+   for (int32_t i = 0; i < getNumTotalCompilationThreads(); i++)
+      {
+      TR::CompilationInfoPerThread *curCompThreadInfoPT = _arrayOfCompilationInfoPerThread[i];
+      if (!curCompThreadInfoPT->isDiagnosticThread())
+         {
+         do
+            {
+            /* Determine whether to wait on the CR Monitor.
+             *
+             * IMPORTANT: There should be no return, or C++ exception thrown, after
+             *            releasing the Comp Monitor below until it is re-acquired.
+             *            The Comp Monitor was acquired in an OMR::CriticalSection
+             *            object; a return, or thrown exception, will destruct this
+             *            object as part leaving the scope, or stack unwinding, which
+             *            will attempt to release the monitor in its destructor.
+             */
+            if (!shouldCheckpointBeInterrupted()
+                && curCompThreadInfoPT->getCompilationThreadState() != COMPTHREAD_SUSPENDED)
+               {
+               /* Acquire the CR monitor */
+               acquireCRMonitor();
+
+               /* Release the Comp Monitor before waiting */
+               releaseCompMonitor(vmThread);
+
+               /* Wait until notified */
+               waitOnCRMonitor();
+
+               /* Release CR Monitor and re-acquire the Comp Monitor */
+               releaseCRMonitor();
+               acquireCompMonitor(vmThread);
+               }
+            else
+               {
+               break;
+               }
+            }
+         while(true);
+         }
+
+      /* Stop cycling through the threads if checkpointing should be interrupted. */
+      if (shouldCheckpointBeInterrupted())
+         break;
+      }
+   }
+
+   }
+
+void TR::CompilationInfo::prepareForRestore()
+   {
+
+   {
+   OMR::CriticalSection resumeCompThreadsForRestore(getCompilationMonitor());
+
+   /* Reset the checkpoint in progress flag. */
+   resetCheckpointInProgress();
+
+   /* Resume suspended compilation threads. */
+   resumeCompilationThread();
+   }
+
+   }
+#endif // #if defined(J9VM_OPT_CRIU_SUPPORT)
 
 int TR::CompilationInfo::computeCompilationThreadPriority(J9JavaVM *vm)
    {
@@ -3407,6 +3516,18 @@ void TR::CompilationInfo::stopCompilationThreads()
 #endif
    acquireCompMonitor(vmThread);
 
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+   // Interrupt any checkpoint in progress and notify
+   // the thread waiting in the checkpoint hook.
+   if (isCheckpointInProgress())
+      {
+      interruptCheckpoint();
+      acquireCRMonitor();
+      getCRMonitor()->notifyAll();
+      releaseCRMonitor();
+      }
+#endif
+
    // Cycle through all non-diagnostic threads and stop them
    for (int32_t i = 0; i < getNumTotalCompilationThreads(); i++)
       {
@@ -3872,13 +3993,34 @@ TR::CompilationInfoPerThread::doSuspend()
    {
    _compInfo.setSuspendThreadDueToLowPhysicalMemory(false);
    getCompThreadMonitor()->enter();
+
+   // Set the new state
    setCompilationThreadState(COMPTHREAD_SUSPENDED);
-   _compInfo.releaseCompMonitor(getCompilationThread());   // release the queue monitor before waiting
+
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+   // Notify the thread waiting in the checkpoint hook.
+   if (_compInfo.isCheckpointInProgress())
+      {
+      _compInfo.acquireCRMonitor();
+      _compInfo.getCRMonitor()->notifyAll();
+      _compInfo.releaseCRMonitor();
+      }
+#endif
+
+   // release the queue monitor before waiting
+   _compInfo.releaseCompMonitor(getCompilationThread());
+
    setLastTimeThreadWasSuspended(_compInfo.getPersistentInfo()->getElapsedTime());
    setVMThreadNameWithFlag(getCompilationThread(), getCompilationThread(), getSuspendedThreadName(), 1);
-   getCompThreadMonitor()->wait(); // wait here until someone notifies us
+
+   // wait here until someone notifies us
+   getCompThreadMonitor()->wait();
+
    setVMThreadNameWithFlag(getCompilationThread(), getCompilationThread(), getActiveThreadName(), 1);
+
    getCompThreadMonitor()->exit();
+
+   // reacquire the queue monitor
    _compInfo.acquireCompMonitor(getCompilationThread());
    }
 
@@ -5807,7 +5949,12 @@ void *TR::CompilationInfo::compileOnSeparateThread(J9VMThread * vmThread, TR::Il
    // end this compilation request. The only exception to this case is if we are performing a JitDump compilation,
    // in which case we always want to proceed with the compilation since it will be performed on the diagnostic
    // thread. Note that the diagnostic thread is not counted towards `getNumCompThreadsActive` count.
-   if ((getNumCompThreadsActive() == 0 || getPersistentInfo()->getDisableFurtherCompilation()) && !details.isJitDumpMethod())
+   if ((getNumCompThreadsActive() == 0
+#if defined(J9VM_OPT_CRIU_SUPPORT)
+        || isCheckpointInProgress()
+#endif
+        || getPersistentInfo()->getDisableFurtherCompilation())
+       && !details.isJitDumpMethod())
       {
       bool shouldReturn = true;
